requirements.txt
asgiref==3.7.2
attrs==23.1.0
cfgv==3.4.0
distlib==0.3.7
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
filelock==3.12.4
identify==2.5.29
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
nodeenv==1.8.0
packaging==23.1
platformdirs==3.10.0
pre-commit==3.4.0
PyJWT==1.7.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
virtualenv==20.24.5
